DjangoPark2023/requirements.txt
asgiref==3.6.0
baidu-aip==4.16.10
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
Django==4.1.7
django-cleanup==7.0.0
idna==3.4
Pillow==9.4.0
PyMySQL==1.0.3
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
